moved Tile folder to layouts folder
